+++ src/protocol.hpp
@@ -384,7 +384,7 @@ void ProtocolInput( descriptor_t *apDescriptor, char *apData, int aSize, char *a
  * Note that the MXP tags will automatically be removed if the user doesn't
  * support MXP, but it's very important you remember to close the tags.
  */
-const char *ProtocolOutput( descriptor_t *apDescriptor, const char *apData, int *apLength, bool appendGA );
+const char *ProtocolOutput( descriptor_t *apDescriptor, const char *apData, size_t *apLength, bool appendGA );
 #define DO_APPEND_GA   1
 #define DONT_APPEND_GA 0
